requirements.txt
chromadb==0.5.3
numpy
pyyaml
PyMuPDF
python-docx
langchain
sentence_transformers
google.generativeai
